refactor: cover the scenario where transfer are sponsored. Change tests
